+++ src/forward/proto/inject_sctp.c
@@ -46,10 +46,11 @@ struct inject_sctp_context_struct{
 	// see more: https://linux.die.net/man/3/sctp_sendmsg
 	sctp_param_t sctp_param;
 	bool shown_error;
+	size_t total_sent_pkt; //total packets being sent successfully
 };
 
 void _sctp_connect( inject_sctp_context_t *context ){
-	int conn_fd, ret;
+	int conn_fd, ret, opt;
 	struct sockaddr_in servaddr = {
 			.sin_family = AF_INET,
 			.sin_port = htons( context->port ),
@@ -59,8 +60,18 @@ void _sctp_connect( inject_sctp_context_t *context ){
 	conn_fd = socket(AF_INET, SOCK_STREAM, IPPROTO_SCTP);
 	ASSERT( conn_fd >= 0, "Cannot create SCTP socket" );
 
-	ret = connect(conn_fd, (struct sockaddr *) &servaddr, sizeof(servaddr));
-	ASSERT( ret >= 0, "Cannot connect to %s:%d using SCTP", context->host, context->port );
+	opt = 1;
+	ret = setsockopt(conn_fd, IPPROTO_SCTP, SCTP_NODELAY, &opt, sizeof(opt));
+	ASSERT( ret >= 0, "Cannot set SCTP_NODELAY to SCTP socket" );
+
+	do {
+		ret = connect(conn_fd, (struct sockaddr *) &servaddr, sizeof(servaddr));
+		if( ret == -1 ){
+			log_write( LOG_ERR, "Cannot connect to %s:%d using SCTP: %s. Try to reconnect in 10 seconds.",
+					context->host, context->port, strerror(errno) );
+			sleep( 10 );
+		}
+	} while( ret == -1 );
 
 	context->client_fd = conn_fd;
 	context->shown_error = false;
@@ -73,6 +84,7 @@ inject_sctp_context_t* inject_sctp_alloc( const forward_packet_target_conf_t *co
 	context->host      = conf->host;
 	context->port      = conf->port;
 	context->nb_copies = nb_copies;
+	context->total_sent_pkt = 0;
 	//sctp parameters to send packets
 	inject_sctp_update_param( context,
 			60, //payload protocol id of NGAP protcol
@@ -90,18 +102,35 @@ void inject_sctp_update_param( inject_sctp_context_t *context, uint32_t ppid, ui
 	context->sctp_param.timetolive = timetolive;
 }
 
-static inline void _reconnect_sctp_if_need( inject_sctp_context_t *context ){
+static inline void _print_sctp_status( inject_sctp_context_t *context ){
 	struct sctp_status status;
+	/*
+	struct sctp_status {
+		sctp_assoc_t      sstat_assoc_id;
+		__s32       sstat_state;
+		__u32       sstat_rwnd;
+		__u16       sstat_unackdata;
+		__u16       sstat_penddata;
+		__u16       sstat_instrms;
+		__u16       sstat_outstrms;
+		__u32       sstat_fragmentation_point;
+		struct sctp_paddrinfo   sstat_primary;
+	};
+	 */
 	int ret;
-    int i = sizeof(status);
-    ret = getsockopt( context->client_fd, SOL_SCTP, SCTP_STATUS, &status, (socklen_t *)&i);
-    ASSERT( ret != 0, "Cannot get SCTP socket status");
-
-    printf("\nSCTP Status:\n--------\n");
-    printf("assoc id  = %d\n", status.sstat_assoc_id);
-    printf("state     = %d\n", status.sstat_state);
-    printf("instrms   = %d\n", status.sstat_instrms);
-    printf("outstrms  = %d\n--------\n\n", status.sstat_outstrms);
+	int i = sizeof(status);
+	//will get error as we need to provide status.sstat_assoc_id which need to be assigned when getting msg from server
+	// https://stackoverflow.com/questions/23897781/getsockopt-invalid-argument-for-ipproto-sctp-sctp-status
+	// ignore this for now
+	ret = getsockopt( context->client_fd, IPPROTO_SCTP, SCTP_STATUS, &status, (socklen_t *)&i);
+	if(ret != 0)
+		log_write( LOG_ERR, "Cannot get SCTP socket status: %s", strerror(errno));
+	else {
+		log_write( LOG_INFO, "SCTP Status: assoc id=%d, state=%d, unackdata=%d, penddata=%d, instrms=%d, outstrms=%d",
+			status.sstat_assoc_id, status.sstat_state,
+			status.sstat_unackdata, status.sstat_penddata,
+			status.sstat_instrms, status.sstat_outstrms);
+	}
 }
 
 static inline void _clear_sctp_buffer_if_need( inject_sctp_context_t *context ){
@@ -121,7 +150,6 @@ int inject_sctp_send_packet( inject_sctp_context_t *context, const uint8_t *pack
 	uint16_t nb_pkt_sent = 0;
 	int ret, i;
 
-//	_reconnect_sctp_if_need( context );
 	_clear_sctp_buffer_if_need( context );
 
 	for( i=0; i<context->nb_copies; i++ ){
@@ -135,12 +163,25 @@ int inject_sctp_send_packet( inject_sctp_context_t *context, const uint8_t *pack
 				context->sctp_param.context   //context
 			);
 		//ret = sctp_send( context->client_fd, packet_data, packet_size, NULL, 0 );
-		if( ret > 0 )
+		if( ret > 0 ){
 			nb_pkt_sent ++;
-		else if(! context->shown_error ){
+			context->total_sent_pkt ++;
+			//clear the reception buffer for each X sending packets
+			if( nb_pkt_sent % 10 == 0 )
+				_clear_sctp_buffer_if_need( context );
+		} else if(! context->shown_error ){
 			context->shown_error = true;
-			log_write( LOG_ERR, "SCTP error when injecting payload %d: %s (please check sctp parameters)", ret, strerror(errno) );
+			log_write( LOG_ERR, "SCTP error when injecting %zu-th packet %d: %s (please check sctp parameters)",
+					context->total_sent_pkt, ret, strerror(errno) );
+			//log current status of the socket
+			//_print_sctp_status( context );
 		}
+		//reconnect if need
+		if( ret == -1 ){
+			close(context->client_fd);
+			_sctp_connect( context );
+		}
+
 	}
 
 	//sleep(1);